Complete part 4 exercises

+++ part4/bloglist/controllers/blogs.js
@@ -2,7 +2,8 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 blogsRouter.get('/:id', async (request, response) => {
@@ -14,8 +15,11 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
+
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
+
   if (typeof body.title === 'undefined' || typeof body.url === 'undefined' ) {
     response.status(400).end()
   } else {
@@ -23,16 +27,32 @@ blogsRouter.post('/', async (request, response) => {
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes
+      likes: body.likes,
+      user: user._id
     })
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
   }
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  let match = false
+  const user = request.user
+  user.blogs.forEach(blog => {
+    if(blog.toString() === request.params.id) {
+      match = true
+    }
+  })
+  if (!match) {
+    response.status(400).json({ error: 'blog does not belong to user' })
+  } else {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
 
 })
